1st-game/pyproject.toml
[project]
name = "star-catcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.6",
]
